Miniboa
Add miniboa.py library and its license file to atria/libs package.

This is a Python 3 port of the Miniboa library. We're adding it to our
repo instead of having it as a build requirement because I'll likely want
to make some changes to it for my implementation.

For more details on the port, see: https://code.google.com/p/miniboa-py3/
For details on the original project, see: https://code.google.com/p/miniboa/