Miniboa Changes
Update library to use our configured logging instead of the default one,
so that it can log to "miniboa" instead of the root logger. Add another
check for dropped clients near the end of TelnetServer.poll so that it
doesn't take two calls to poll to find out someone dropped.