weak_property Decorator Rework
Change the decorated function passed to _make_weak_property from a postsetter
to a validator (it now takes two arguments, the old and new objects, and will
be called prior to the new weak reference being created and set).

I think I'm going to change this decorator around completely to be a mix-in
class with a series of methods for managing weak references instead of a
decorator, that will give a lot more control over validating and pre or post
settings, etc. Besides, the decorator is pretty ugly when you have a bunch
of empty decorated functions just to make the property.