Shell Modules
Add shells.py module to the atria/core package, along with tests.

These are pretty barebones at the moment, until command processing
can be integrated. I'm going to rewrite the argument parsing to be a
little smarter and not use regular expressions, but I figured I'd commit
this first for posterity.