Server State Saving
Add save_state and load_state functions to server.py module. Update boot
function to look for an existing state file and load it if found.

The state system at this point is just a framework, individual modules
should hook the server_save_state and server_load_state events to push in
and pull out their data for processing.