Session State Saving
Add hooks to server_save_state and server_load_state events in sessions.py
module to save and load session data, tying back to a client if available.

With this commit, I was able to spawn two servers, connect to one over
telnet and have that connection passed to the other server without
interruption, so old ROM-style copyover is certainly within reach.
(I tested it in Windows which is the harder platform to get it working on,
getting it up and running in Linux should be trivial.)