Server Reload Logic
Add handling of a -R command line argument to the __main__.py module
that tells the server it is running in reload mode and should wait
for a server state file to be available before booting. Add a
ServerReload exception to the server.py module. Add handling for the
ServerReload exception to the loop function.

When the reload exception is raised, a new server process will be
spawned, the server state will be dumped to file and all sockets will
be passed to the new process. In my tests I was able to reload the
server state without interruption to connected clients and everything
worked perfectly, however there were some quirky issues with dropping
back to the command line after the initial process died while still
receiving log output from the new process in the original window.
This only seemed to happen when python was called directly in the
command prompt with "python -m atria". I created a batch file that
ran the same command and it does not drop back to the command line
after a reload. Further testing will be needed (still need to test
on Linux as well, this was all done in Windows, which I figured
would be the quirkier of the two).