HasParent Mix-in Rework
Move all of the HasParent mix-in's functionality to a HasParentMeta
metaclass that HasParent derives from.

Now the parent system works only on classes themselves and not their
instances. I can't currently think of a use-case for individual instances
having their own unique parent fields but if there is one, the system
can be extended.