HasFlags and HasWeaks Mix-ins Rework
Duplicate functionality of the HasFlags and HasWeaks mix-ins into the
HasFlagsMeta and HasWeaksMeta metaclasses which they each derive from,
respectively.

Using a metaclass system allows both classes and instances of classes
to store flags and weak data individually without instances conflicting
with the data on their classes. However, this complicates the mix-in
system a bit when using multiple mix-ins together as multi-meta base
classes will need to be constructed.