Request System
Expand functionality of Request class in atria/core/requests.py module.

This new system is much simpler compared to those I cooked up in previous
iterations of the MUD; resolve simply returns True or False depending on
whether the a request gets resolved or not by the given input. This way
the handler in sessions doesn't need to worry about handling different
exceptions based on the outcome as was used in my previous system.