Entity Version
Add EntityVersion Attribute and register "version" on the base Entity class
so that it is inherited by all subclass entities.

Entity versions can be used by the load and save methods of subclasses to
do any necessary updating on an entity that hasn't been loaded since before
a change was made to its data structure, etc.