Entity Instance Caching
Add an _instances attribute with a WeakValueDictionary unique to each Entity
classe via the _EntityMeta init method. Add tracking of entity instances
through the _instances dict via the Entity init method. Add from_cache
keyword argument to Entity.load; if True, it will check the _instances cache
before hitting the store. Change Entity.load to not require an Entity class
to have a store set, it will instead just fall down to the KeyError when no
entity is found. Change Entity.load to warn when a loaded entity did not have
a UID in it's data instead of raising an exception (it's not a big deal
anyway, a new UID would be assigned then the entity is instantiated and that
UID would be saved the next time the entity is, but it's good to know that
it happened).