Entity UID Handling
Move manual handling of UIDs from the load, save, and revert methods of
Entity to the serialize and deserialize methods. Change the init method of
Entity to only set a new UID if one was not deserialized. Change the
FileNotFoundError exception raised in Entity.load to a KeyError instead.

Entity UIDs should now only be set in the init, deserialize, and
(potentially, if the _store_key is "uid") clone methods. You should NOT
be manually setting the _uid attribute anywhere else.