Entity.find Method
Add find method to Entity for searching for entities by the value of a
particular attribute.

Entity.find should be used pretty sparingly when dealing with disk-based
stores and searching the store. For instance finding an entity that has
a PickleStore will require reading through every pickle file in the store.
This should later be abstracted to a _find method that is overridden by each
store type so that they can be as efficient as possible with their searches.