Miniboa TelnetServer Updates
Add __repr__ method to TelnetClient. Add server_socket keyword argument to
the TelnetServer constructor; if server_socket is is passed and is not 0,
the provided socket is used to listen instead of creating a new one, if it
is 0, no listner is created. Add create_client keyword argument to
TelnetServer constructor; if True, new clients are created from connecting
sockets and those clients are passed to the on_connect callback (the default
and previous behavior), if False, the server does not create a client for
new sockets and instead passes the socket and address tuple to the
on_connect callback.