Miniboa Restyled
Bring miniboa.py in line with PEP8, add docstrings where missing, and
normalize others with the rest of the project. Remove example chat server
from the bottom of the library. Remove now-superfluous setup.cfg file.

I wouldn't normally do this to a third-party library but as I've already
made some changes to it and will probably hack it up even further, I thought
I'd get it in line with the rest of the project style.