Entity.active
Add 'active' attribute to Entity to represent whether an entity is currently
"in play" or not (it currently has no effect on anything). Add 'resume' and
'suspend' stub methods to Character to handle bringing characters in and out
of the game world (I may move this functionality to Entity itself and let
subclasses extend a base method, but this is a jumping off point).