UID Generation Overhaul
Remove make_time_codes generator from utils/funcs.py module and replace calls
to it with int_to_base_n. Move any UID-related code (timecode generation) from
timing.py module and consolidate it in entities.py module. Add most capital
letters to the timecode character set (and rename the character set to
_uid_timecode_charset). Add _uid_timecode_multiplier setting to entities.py
(it can be tweaked to change timecode precision). Remove all the UID history
tracking stuff as it is unnecessary with new system. Rewrite Entity.make_uid
to handle all of the UID creation logic. Remove an obsolete test.

Instead of counting UIDs of a certain type each pulse, the new system will
increment the current time and generate a new timecode. This feels better to
me. Not only will a timecode be unique regardless of entity type, but we
don't need the "-0" that would have been appended to 99% of UIDs. This does
however introduce the possibility of collision, in the event that you create
a very large amount of UIDs in one pulse, pushing the timecode so far past
the system time that a server restart would roll back the timecode. (This is
very, very unlikely though and maybe borders on impossible, as you'd have to
create more than 10,000 Entities per second.)