Entity Store Keys
Remove individual stores and store keys from entity subclasses. Add central
JSON store to Entity. Remove `key` property and `get_key_name` method from
Entity, and change all references of Entity.key to .uid.

(The entity system is in a state of transition at the moment, and the engine
won't currently compile as it is.)