Temperature Layer
Add a temperature layer to the world generation mechanics. Rewrite the
terrain grid file to be three dimensional, to account for the new layer.
Add a bunch of new terrain types to cover the cold and moderate parts
of the temperature spectrum (still need to add more for the hot part).